+++ src/string.cc
@@ -82,28 +82,28 @@ namespace kortex {
     }
 
     string num2str(const float& m, const int& precision) {
-        float  n      = fabs(m);
+        float  n      = std::fabs(m);
         string retval = num2str(int(n));
         retval += ".";
-        n -= int(n);
+        n -= float( int(n) );
         for( int d=0; d<precision; d++ ) {
             n = 10*n;
             retval += num2str(int(n));
-            n -= int(n);
+            n -= float( int(n) );
         }
         if( m<0 ) retval="-"+retval;
         return retval;
     }
 
     string num2str(const double& m, const int& precision) {
-        double n = fabs(m);
+        double n = std::fabs(m);
         string retval = num2str(int(n));
         retval += ".";
-        n -= int(n);
+        n -= double( int(n) );
         for( int d=0; d<precision; d++ ) {
             n = 10*n;
             retval += num2str(int(n));
-            n -= int(n);
+            n -= double( int(n) );
         }
         if( m<0 ) retval="-"+retval;
         return retval;
